+++ ulid/_util.ts
@@ -1,6 +1,12 @@
 // Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
 
-import type { PRNG } from "./mod.ts";
+interface ULID {
+  (seedTime?: number): string;
+}
+
+interface PRNG {
+  (): number;
+}
 
 // These values should NEVER change. If
 // they do, we're no longer making ulids!
@@ -10,11 +16,11 @@ export const TIME_MAX = Math.pow(2, 48) - 1;
 export const TIME_LEN = 10;
 export const RANDOM_LEN = 16;
 
-export function replaceCharAt(str: string, index: number, char: string) {
+function replaceCharAt(str: string, index: number, char: string) {
   return str.substring(0, index) + char + str.substring(index + 1);
 }
 
-export function randomChar(prng: PRNG): string {
+function randomChar(prng: PRNG): string {
   let rand = Math.floor(prng() * ENCODING_LEN);
   if (rand === ENCODING_LEN) {
     rand = ENCODING_LEN - 1;
@@ -22,7 +28,7 @@ export function randomChar(prng: PRNG): string {
   return ENCODING.charAt(rand);
 }
 
-export function encodeTime(now: number, len: number = TIME_LEN): string {
+function encodeTime(now: number, len: number = TIME_LEN): string {
   if (now > TIME_MAX) {
     throw new Error("cannot encode time greater than " + TIME_MAX);
   }
@@ -41,7 +47,7 @@ export function encodeTime(now: number, len: number = TIME_LEN): string {
   return str;
 }
 
-export function encodeRandom(len: number, prng: PRNG): string {
+function encodeRandom(len: number, prng: PRNG): string {
   let str = "";
   for (; len > 0; len--) {
     str = randomChar(prng) + str;
@@ -49,7 +55,7 @@ export function encodeRandom(len: number, prng: PRNG): string {
   return str;
 }
 
-export function detectPrng(): PRNG {
+function detectPrng(): PRNG {
   return () => {
     const buffer = new Uint8Array(1);
     crypto.getRandomValues(buffer);
@@ -57,7 +63,7 @@ export function detectPrng(): PRNG {
   };
 }
 
-export function incrementBase32(str: string): string {
+function incrementBase32(str: string): string {
   let index = str.length;
   let char;
   let charIndex;
@@ -75,4 +81,64 @@ export function incrementBase32(str: string): string {
     return replaceCharAt(str, index, ENCODING[charIndex + 1]);
   }
   throw new Error("cannot increment this string");
+}
+
+/**
+ * Generates a ULID function given a PRNG
+ *
+ * @example To use your own pseudo-random number generator, import the factory, and pass it your generator function.
+ * ```ts
+ * import { factory } from "https://deno.land/std@$STD_VERSION/ulid/mod.ts";
+ * const prng = () => Math.random();
+ *
+ * const ulid = factory(prng);
+ * ulid(); // 01BXAVRG61YJ5YSBRM51702F6M
+ * ```
+ */
+export function factory(prng: PRNG = detectPrng()): ULID {
+  return function ulid(seedTime: number = Date.now()): string {
+    return encodeTime(seedTime, TIME_LEN) + encodeRandom(RANDOM_LEN, prng);
+  };
+}
+
+/**
+ * Generates a monotonically increasing ULID, optionally given a PRNG.
+ *
+ * @example To generate monotonically increasing ULIDs, create a monotonic counter.
+ * ```ts
+ * import { monotonicFactory } from "https://deno.land/std@$STD_VERSION/ulid/mod.ts";
+ *
+ * const ulid = monotonicFactory();
+ * // Strict ordering for the same timestamp, by incrementing the least-significant random bit by 1
+ * ulid(150000); // 000XAL6S41ACTAV9WEVGEMMVR8
+ * ulid(150000); // 000XAL6S41ACTAV9WEVGEMMVR9
+ * ulid(150000); // 000XAL6S41ACTAV9WEVGEMMVRA
+ * ulid(150000); // 000XAL6S41ACTAV9WEVGEMMVRB
+ * ulid(150000); // 000XAL6S41ACTAV9WEVGEMMVRC
+ *
+ * // Even if a lower timestamp is passed (or generated), it will preserve sort order
+ * ulid(100000); // 000XAL6S41ACTAV9WEVGEMMVRD
+ * ```
+ *
+ * @example You can also pass in a prng to the monotonicFactory function.
+ * ```ts
+ * import { monotonicFactory } from "https://deno.land/std@$STD_VERSION/ulid/mod.ts";
+ * const prng = () => Math.random();
+ *
+ * const ulid = monotonicFactory(prng);
+ * ulid(); // 01BXAVRG61YJ5YSBRM51702F6M
+ * ```
+ */
+export function monotonicFactory(prng: PRNG = detectPrng()): ULID {
+  let lastTime = 0;
+  let lastRandom: string;
+  return function ulid(seedTime: number = Date.now()): string {
+    if (seedTime <= lastTime) {
+      const incrementedRandom = (lastRandom = incrementBase32(lastRandom));
+      return encodeTime(lastTime, TIME_LEN) + incrementedRandom;
+    }
+    lastTime = seedTime;
+    const newRandom = (lastRandom = encodeRandom(RANDOM_LEN, prng));
+    return encodeTime(seedTime, TIME_LEN) + newRandom;
+  };
 }
