+++ ulid/mod.ts
@@ -13,24 +13,14 @@
  */
 
 import {
-  detectPrng,
-  encodeRandom,
-  encodeTime,
   ENCODING,
   ENCODING_LEN,
-  incrementBase32,
+  factory,
+  monotonicFactory,
   RANDOM_LEN,
   TIME_LEN,
   TIME_MAX,
 } from "./_util.ts";
-
-export interface PRNG {
-  (): number;
-}
-
-export interface ULID {
-  (seedTime?: number): string;
-}
 
 /**
  * Extracts the timestamp given a ULID
@@ -57,64 +47,22 @@ export function decodeTime(id: string): number {
 }
 
 /**
- * Generates a ULID function given a PRNG
- *
- * @example To use your own pseudo-random number generator, import the factory, and pass it your generator function.
+ * @example
  * ```ts
- * import { factory } from "https://deno.land/std@$STD_VERSION/ulid/mod.ts";
- * const prng = () => Math.random();
+ * import { monotonicUlid } from "https://deno.land/std@$STD_VERSION/ulid/mod.ts";
  *
- * const ulid = factory(prng);
- * ulid(); // 01BXAVRG61YJ5YSBRM51702F6M
- * ```
- */
-export function factory(prng: PRNG = detectPrng()): ULID {
-  return function ulid(seedTime: number = Date.now()): string {
-    return encodeTime(seedTime, TIME_LEN) + encodeRandom(RANDOM_LEN, prng);
-  };
-}
-
-/**
- * Generates a monotonically increasing ULID, optionally given a PRNG.
- *
- * @example To generate monotonically increasing ULIDs, create a monotonic counter.
- * ```ts
- * import { monotonicFactory } from "https://deno.land/std@$STD_VERSION/ulid/mod.ts";
- *
- * const ulid = monotonicFactory();
  * // Strict ordering for the same timestamp, by incrementing the least-significant random bit by 1
- * ulid(150000); // 000XAL6S41ACTAV9WEVGEMMVR8
- * ulid(150000); // 000XAL6S41ACTAV9WEVGEMMVR9
- * ulid(150000); // 000XAL6S41ACTAV9WEVGEMMVRA
- * ulid(150000); // 000XAL6S41ACTAV9WEVGEMMVRB
- * ulid(150000); // 000XAL6S41ACTAV9WEVGEMMVRC
+ * monotonicUlid(150000); // 000XAL6S41ACTAV9WEVGEMMVR8
+ * monotonicUlid(150000); // 000XAL6S41ACTAV9WEVGEMMVR9
+ * monotonicUlid(150000); // 000XAL6S41ACTAV9WEVGEMMVRA
+ * monotonicUlid(150000); // 000XAL6S41ACTAV9WEVGEMMVRB
+ * monotonicUlid(150000); // 000XAL6S41ACTAV9WEVGEMMVRC
  *
  * // Even if a lower timestamp is passed (or generated), it will preserve sort order
- * ulid(100000); // 000XAL6S41ACTAV9WEVGEMMVRD
- * ```
- *
- * @example You can also pass in a prng to the monotonicFactory function.
- * ```ts
- * import { monotonicFactory } from "https://deno.land/std@$STD_VERSION/ulid/mod.ts";
- * const prng = () => Math.random();
- *
- * const ulid = monotonicFactory(prng);
- * ulid(); // 01BXAVRG61YJ5YSBRM51702F6M
+ * monotonicUlid(100000); // 000XAL6S41ACTAV9WEVGEMMVRD
  * ```
  */
-export function monotonicFactory(prng: PRNG = detectPrng()): ULID {
-  let lastTime = 0;
-  let lastRandom: string;
-  return function ulid(seedTime: number = Date.now()): string {
-    if (seedTime <= lastTime) {
-      const incrementedRandom = (lastRandom = incrementBase32(lastRandom));
-      return encodeTime(lastTime, TIME_LEN) + incrementedRandom;
-    }
-    lastTime = seedTime;
-    const newRandom = (lastRandom = encodeRandom(RANDOM_LEN, prng));
-    return encodeTime(seedTime, TIME_LEN) + newRandom;
-  };
-}
+export const monotonicUlid = monotonicFactory();
 
 /**
  * @example
